requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.21.6
requests==2.31.0

# Data Processing & Analysis
scikit-learn==1.0.2
tensorflow==2.11.0
ta==0.10.2
scipy==1.10.1
statsmodels==0.13.5

# Trading & Exchange APIs
ccxt==4.0.0
python-binance==1.0.16
kucoin-python==2.1.3

# Database
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
pymongo==4.3.3

# Visualization
matplotlib==3.5.3
seaborn==0.12.2
plotly==5.13.0

# Web Framework & API
fastapi==0.95.0
uvicorn==0.21.1
websockets==10.4

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0

# Development Tools
black==23.3.0
flake8==6.0.0
mypy==1.1.1
isort==5.12.0

# Logging & Monitoring
structlog==23.1.0
sentry-sdk==1.17.0

# Task Queue & Scheduling
celery==5.2.7
redis==4.5.4
apscheduler==3.10.1

# Documentation
sphinx==6.1.3
sphinx-rtd-theme==1.2.0

# Security
cryptography==40.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt